% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_sol_tbl.R
\name{eq_sol_tbl}
\alias{eq_sol_tbl}
\title{Generate a lead solubility table using inputs from \code{shiny}.}
\usage{
eq_sol_tbl(input, model = "leadsol", dbase = pbcusol:::pbcu2sol)
}
\arguments{
\item{input}{Input from \code{shiny} UI.}

\item{model}{Currently either "leadsol" or "minteq.v4".}

\item{dbase}{A \code{phreeqc} or \code{pbcusol} database.}
}
\value{
A numeric vector of lead concentrations.
}
\description{
Generate a lead solubility table using inputs from \code{shiny}.
}
