% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypbsol.R
\name{shinypbsol}
\alias{shinypbsol}
\title{Calculate equilibrium lead solubility using \code{pbcusol} and \code{shiny}}
\usage{
shinypbsol(...)
}
\arguments{
\item{...}{Argument passed to function. (Currently ignored.)}
}
\description{
This function is used in deploying the \code{shinypbcusol} \code{shiny} app.
}
