% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kinetics.R
\name{plot_kinetics}
\alias{plot_kinetics}
\title{Plot a kinetic metal release curve due to stagnation in a pipe.}
\usage{
plot_kinetics(eq_sol)
}
\arguments{
\item{eq_sol}{The equilibrium solubility value.}
}
\value{
A ggplot2 object.
}
\description{
Plot a kinetic metal release curve due to stagnation in a pipe.
}
\examples{
plot_kinetics(100)
}
